Make CreateNote component
